+++ types/Version.d.ts
@@ -1,4 +1,4 @@
 /**
  * Constant with the hardcoded version of the app
  */
-export declare const VERSION = "1.1.0";
+export declare const VERSION = "1.1.1";